% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug_typed-package.R
\docType{package}
\name{Bug_typed-package}
\alias{Bug_typed}
\alias{Bug_typed-package}
\title{Bug_typed: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
